+++ packages/deposit-screener/scripts/genSignedDepositRequestTestCase.ts
@@ -1,15 +1,15 @@
-import { TypedDataSigner, Signer } from "@ethersproject/abstract-signer";
 import { AssetTrait, AssetType, DepositRequest } from "@nocturne-xyz/sdk";
 import { ERC20_ID } from "@nocturne-xyz/sdk/dist/src/primitives/asset";
-import { Wallet } from "ethers";
+import { ethers, Wallet } from "ethers";
 import { signDepositRequest } from "../src";
-import * as JSON from "bigint-json-serialization";
 import findWorkspaceRoot from "find-yarn-workspace-root";
 import * as path from "path";
 import * as fs from "fs";
 import {
   DEPOSIT_CHECKER_CONTRACT_NAME,
   DEPOSIT_CHECKER_CONTRACT_VERSION,
+  DEPOSIT_REQUEST_TYPES,
+  EIP712Domain,
 } from "../src/typedData";
 
 const ROOT_DIR = findWorkspaceRoot()!;
@@ -20,41 +20,72 @@ const SIGNED_DEPOSIT_REQ_FIXTURE_PATH = path.join(
 
 const writeToFixture = process.argv[2] == "--writeFixture";
 
+function toObject(obj: any) {
+  return JSON.parse(
+    JSON.stringify(obj, (key, value) =>
+      typeof value === "bigint" ? value.toString() : value
+    )
+  );
+}
+
 (async () => {
-  const depositCheckerAddress = "0x1111111111111111111111111111111111111111";
-  const signer: Signer & TypedDataSigner = new Wallet(
+  const signer = new Wallet(
     "0x0000000000000000000000000000000000000000000000000000000000000001"
   );
+
+  const depositManagerAddress = "0x1111111111111111111111111111111111111111";
+  const chainId = 123n;
+  const domain: EIP712Domain = {
+    name: DEPOSIT_CHECKER_CONTRACT_NAME,
+    version: DEPOSIT_CHECKER_CONTRACT_VERSION,
+    chainId,
+    verifyingContract: depositManagerAddress,
+  };
+
+  const encodedAsset = AssetTrait.encode({
+    assetType: AssetType.ERC20,
+    assetAddr: "0x0000000000000000000000000000000000000123",
+    id: ERC20_ID,
+  });
+
   const depositRequest: DepositRequest = {
-    chainId: 123n,
+    chainId,
     spender: await signer.getAddress(),
-    encodedAsset: AssetTrait.encode({
-      assetType: AssetType.ERC20,
-      assetAddr: "0x0000000000000000000000000000000000000123",
-      id: ERC20_ID,
-    }),
+    encodedAsset,
     value: 1000n,
     depositAddr: {
       h1X: 1n,
       h1Y: 2n,
       h2X: 3n,
-      h2Y: 4n,
+      h2Y: 5n,
     },
     nonce: 0n,
     gasPrice: 50n,
   };
 
-  const signedDepositRequest = await signDepositRequest(
-    signer,
-    depositRequest,
-    depositCheckerAddress
-  );
+  const signature = await signDepositRequest(signer, domain, depositRequest);
+  const { r, s, v } = ethers.utils.splitSignature(signature);
 
-  const json = JSON.stringify({
-    depositCheckerContractName: DEPOSIT_CHECKER_CONTRACT_NAME,
-    depositCheckerContractVersion: DEPOSIT_CHECKER_CONTRACT_VERSION,
-    signedDepositRequest,
-  });
+  const expectedSignerAddress = await signer.getAddress();
+  const recoveredAddress = ethers.utils.verifyTypedData(
+    domain,
+    DEPOSIT_REQUEST_TYPES,
+    depositRequest,
+    signature
+  );
+  console.log("Recovered:", recoveredAddress);
+  console.log("Actual:", expectedSignerAddress);
+
+  const json = JSON.stringify(
+    toObject({
+      contractAddress: depositManagerAddress,
+      contractName: DEPOSIT_CHECKER_CONTRACT_NAME,
+      contractVersion: DEPOSIT_CHECKER_CONTRACT_VERSION,
+      screenerAddress: await signer.getAddress(),
+      depositRequest,
+      signature: { r, s, v },
+    })
+  );
   console.log(json);
 
   if (writeToFixture) {